Refactoring.

+++ adpay/utils/common.py
@@ -1,6 +1,7 @@
 from adpay.stats import consts as stats_consts
 import math
 
+
 def timestamp2hour(timestamp):
     """
     Get epoch timestamp of the hour of the `timestamp`.
@@ -8,7 +9,7 @@ def timestamp2hour(timestamp):
     :param timestamp: timestamp
     :return: timestamp of the hour (floored)
     """
-    return math.floor(1.0*timestamp/stats_consts.SECONDS_PER_HOUR)*stats_consts.SECONDS_PER_HOUR
+    return int(math.floor(1.0*timestamp/stats_consts.SECONDS_PER_HOUR)*stats_consts.SECONDS_PER_HOUR)
 
 
 def genkey(key, val, delimiter="_"):
